pyproject.toml
[project]
name = "python-taskgroup-ai-agent"
version = "0.1.0"
description = "AI agent framework using Python's asyncio TaskGroup for concurrent task execution"
readme = "README.md"
authors = [
    { name = "Daku-on", email = "113434878+Daku-on@users.noreply.github.com" }
]
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9.0",
    "asyncpg>=0.30.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.42",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = false
warn_unused_ignores = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false

# Ignore missing imports for third-party libraries
[[tool.mypy.overrides]]
module = [
    "asyncpg.*",
    "aiohttp.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["uv_build>=0.8.0,<0.9"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "ruff>=0.12.7",
]
